Split into library/example
Use tsc instead of vite to build because of issues... :shrug:
