+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "js/examples/cli/webbluetooth"]
-	path = js/examples/cli/webbluetooth
+	path = js/genki-wave/webbluetooth
 	url = git@github.com:genkiinstruments/webbluetooth.git
