+++ js/genki-wave/tsconfig.json
@@ -1,11 +1,17 @@
 {
-  "extends": "../tsconfig.json",
   "compilerOptions": {
-    "module": "commonjs",
     "target": "es5",
-    "sourceMap": true
+    "alwaysStrict": true,
+    "downlevelIteration": true,
+    "noImplicitReturns": true,
+    "noEmitOnError": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "sourceMap": true,
+    "declaration": true,
+    "outDir": "dist"
   },
-  "exclude": [
-    "node_modules"
+  "include": [
+    "src"
   ]
 }
