--- js/genki-wave/vite.config.ts
@@ -1,16 +1,0 @@
-// vite.config.ts
-import { resolve } from 'path'
-import { defineConfig } from 'vite'
-import dts from 'vite-plugin-dts'
-
-// https://vitejs.dev/guide/build.html#library-mode
-export default defineConfig({
-	build: {
-		lib: {
-			entry: resolve(__dirname, 'src/index.ts'),
-			name: 'genki-wave',
-			fileName: 'genki-wave',
-		},
-	},
-	plugins: [dts()],
-})
